mv layout logic to helper
